+++ mod-account/server/authn/.gitignore
@@ -1,0 +1,3 @@
+vendor
+.idea
+
