+++ mod-account/server/authn/go.mod
@@ -1,0 +1,5 @@
+module github.com/getcouragenow/getcourage-packages/mod-account/authn
+
+go 1.13
+
+require google.golang.org/protobuf v1.21.0
