+++ mod-account/server/authn/pkg/auth/errors.go
@@ -1,0 +1,1 @@
+package auth
