+++ mod-account/server/authn/pkg/auth/handler.go
@@ -1,0 +1,1 @@
+package auth
