+++ mod-account/server/keyutils/go.mod
@@ -1,0 +1,5 @@
+module github.com/getcouragenow/getcourage-packages/mod-account/server/keyutils
+
+go 1.13
+
+require github.com/square/go-jose/v3 v3.0.0-20200415055503-21f2ca25ccce
